Set default analysis timeout for all languages
