--- src/Core.UnitTests/CFamily/CFamilyAnalyzerOptionsTests.cs
@@ -1,50 +1,0 @@
-﻿/*
- * SonarLint for Visual Studio
- * Copyright (C) 2016-2020 SonarSource SA
- * mailto:info AT sonarsource DOT com
- *
- * This program is free software; you can redistribute it and/or
- * modify it under the terms of the GNU Lesser General Public
- * License as published by the Free Software Foundation; either
- * version 3 of the License, or (at your option) any later version.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
- * Lesser General Public License for more details.
- *
- * You should have received a copy of the GNU Lesser General Public License
- * along with this program; if not, write to the Free Software Foundation,
- * Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
- */
-
-using FluentAssertions;
-using Microsoft.VisualStudio.TestTools.UnitTesting;
-using Moq;
-using SonarLint.VisualStudio.Core.CFamily;
-
-namespace SonarLint.VisualStudio.Core.UnitTests.CFamily
-{
-    [TestClass]
-    public class CFamilyAnalyzerOptionsTests
-    {
-        [TestMethod]
-        [DataRow(-1, CFamilyAnalyzerOptions.DefaultAnalysisTimeoutMs)]
-        [DataRow(0, CFamilyAnalyzerOptions.DefaultAnalysisTimeoutMs)]
-        [DataRow(1, 1)]
-        [DataRow(999, 999)]
-        public void AnalysisTimeout(int envSettingsResponse, int expectedTimeout)
-        {
-            var envSettingsMock = new Mock<IEnvironmentSettings>();
-            envSettingsMock.Setup(x => x.CFamilyAnalysisTimeoutInMs()).Returns(envSettingsResponse);
-
-            new CFamilyAnalyzerOptions(envSettingsMock.Object).AnalysisTimeoutInMilliseconds.Should().Be(expectedTimeout);
-        }
-
-        [TestMethod]
-        public void AnalysisTimeoutInMilliseconds_NoEnvironmentSettings_DefaultTimeout()
-        {
-            new CFamilyAnalyzerOptions().AnalysisTimeoutInMilliseconds.Should().Be(CFamilyAnalyzerOptions.DefaultAnalysisTimeoutMs);
-        }
-    }
-}
