+++ src/Integration.Vsix.UnitTests/SonarLintTagger/TaggerProviderTests.cs
@@ -31,6 +31,7 @@ using Microsoft.VisualStudio.Text.Editor;
 using Microsoft.VisualStudio.Text.Tagging;
 using Microsoft.VisualStudio.Utilities;
 using Moq;
+using SonarLint.VisualStudio.Core;
 using SonarLint.VisualStudio.Core.Analysis;
 using SonarLint.VisualStudio.Core.Suppression;
 using SonarLint.VisualStudio.Integration.Vsix;
@@ -269,6 +270,25 @@ namespace SonarLint.VisualStudio.Integration.UnitTests
                 });
 
             actual.Should().BeEquivalentTo(trackers);
+        }
+
+        [TestMethod]
+        [DataRow(-1, TaggerProvider.DefaultAnalysisTimeoutMs)]
+        [DataRow(0, TaggerProvider.DefaultAnalysisTimeoutMs)]
+        [DataRow(1, 1)]
+        [DataRow(999, 999)]
+        public void AnalysisTimeout(int envSettingsResponse, int expectedTimeout)
+        {
+            var envSettingsMock = new Mock<IEnvironmentSettings>();
+            envSettingsMock.Setup(x => x.AnalysisTimeoutInMs()).Returns(envSettingsResponse);
+
+            TaggerProvider.GetAnalysisTimeoutInMilliseconds(envSettingsMock.Object).Should().Be(expectedTimeout);
+        }
+
+        [TestMethod]
+        public void AnalysisTimeoutInMilliseconds_NoEnvironmentSettings_DefaultTimeout()
+        {
+            TaggerProvider.GetAnalysisTimeoutInMilliseconds().Should().Be(TaggerProvider.DefaultAnalysisTimeoutMs);
         }
 
         private IIssueTracker[] CreateMockedIssueTrackers(params string[] filePaths) =>
